# GH-45611: [C++][Acero] Improve Swiss join build performance by partitioning batches ahead to reduce contention