--- a/cpp/src/arrow/acero/swiss_join.cc
+++ b/cpp/src/arrow/acero/swiss_join.cc
@@ -1159,7 +1159,7 @@ Status SwissTableForJoinBuild::PartitionBatch(size_t thread_id, int64_t batch_id
                                               const ExecBatch& key_batch,
                                               arrow::util::TempVectorStack* temp_stack) {
   DCHECK_LT(static_cast<int64_t>(thread_id), dop_);
-  DCHECK_LE(batch_id, static_cast<int64_t>(batch_states_.size()));
+  DCHECK_LT(batch_id, static_cast<int64_t>(batch_states_.size()));
   ThreadState& locals = thread_states_[thread_id];
   BatchState& batch_state = batch_states_[batch_id];
   uint16_t num_rows = static_cast<uint16_t>(key_batch.length);