@zanmato1984 2025-02-24
I'm pretty excited by this optimization. @pitrou @westonpace would you mind to take a look? Thanks.