@pitrou 2025-02-24
Does this also remove the spin locks?