@zanmato1984 2025-02-25
> Does this also remove the spin locks?

This removes the spin lock usage in Swiss join build, but the spin lock itself, namely `PartitionLocks::AcquirePartitionLock`, is not removed because it is also used elsewhere.

> Besides the two comments above, I don't feel competent to review this, sorry.

No problem, I appreciate the comments anyway!