@zanmato1984 2025-03-04
The CI failures and regressions are not related.

The benchmarking in my PR description (namely `BM_HashJoinBasic_BuildParallelism`) is not ran because it's disabled by default. But from other relevant benchmark results, namely `BM_HashJoinBasic_ProbeParallelism`, we can see nice performance up when the number of threads is high, for [example](https://conbench.ursa.dev/benchmark-results/067c675a5b8270548000487099f1d9da/) with `512k` rows hash table and `12` threads: 

<img width="834" alt="image" src="https://github.com/user-attachments/assets/b494095f-2eba-404e-b8b4-cd3d7a588a15" />

I'll merge this. And much appreciate the review from @kou and @pitrou !